+++ src/Language/Futhark/Interpreter.hs
@@ -929,7 +929,7 @@ evalDec env (ImportDec name name' loc) =
 
 evalDec env (LocalDec d _) = evalDec env d
 evalDec env SigDec{} = return env
-evalDec env (TypeDec (TypeBind v ps t _ _)) = do
+evalDec env (TypeDec (TypeBind v _ ps t _ _)) = do
   let abbr = T.TypeAbbr Lifted ps $
              evalType env $ unInfo $ expandedType t
   return env { envType = M.insert v abbr $ envType env }